check for DOTFILES_REPO_DIR environment variable before messing things up

+++ scripts/unix/git.sh
@@ -1,4 +1,10 @@
 # link or relink git
+
+if [[ "${DOTFILES_REPO_DIR}" == "" ]]; then
+    echo "DOTFILES_REPO_DIR is not set. Exiting to stay safe"
+    exit 1
+fi
+
 rm -r ~/.gitconfig
 rm -r ~/.gitignore_global
